# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for ForexFactory calendar data, live positions, trade history, and analytics.",
        "title": "Forex Trading Server API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/events": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calendar"
                ],
                "summary": "List calendar events",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum number of events",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter events with event_date greater than or equal to this ISO8601 timestamp",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter events with event_date less than or equal to this ISO8601 timestamp",
                        "name": "date_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.CalendarEvent"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/events/sync": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calendar"
                ],
                "summary": "Trigger a calendar sync",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/report/{login}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trading"
                ],
                "summary": "Get performance report by login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User login",
                        "name": "login",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum history window",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.UserPerformanceReport"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/trading-data": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trading"
                ],
                "summary": "Handle trading data (positions or history)",
                "parameters": [
                    {
                        "description": "Trading data payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.TradingDataRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "List users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum number of users",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.User"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/{user_id}/positions": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trading"
                ],
                "summary": "Upsert a live position snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Position payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.PositionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/{user_id}/report": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trading"
                ],
                "summary": "Generate a performance report for a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum history window",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.UserPerformanceReport"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.CalendarEvent": {
            "type": "object",
            "properties": {
                "actual": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "event": {
                    "type": "string"
                },
                "forecast": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "impact": {
                    "type": "string"
                },
                "previous": {
                    "type": "string"
                },
                "sourceURL": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "domain.UserPerformanceReport": {
            "type": "object",
            "properties": {
                "asOf": {
                    "type": "string"
                },
                "balanceCurve": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.EquityPoint"
                    }
                },
                "closedPositions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.UserTrade"
                    }
                },
                "derivedInsights": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "equityCurve": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.EquityPoint"
                    }
                },
                "exposureMetrics": {
                    "$ref": "#/definitions/domain.ExposureMetrics"
                },
                "openPositions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.UserPositionSnapshot"
                    }
                },
                "realizedProfit": {
                    "type": "number",
                    "format": "float64"
                },
                "riskMetrics": {
                    "$ref": "#/definitions/domain.RiskMetrics"
                },
                "totalReturnPercent": {
                    "type": "number",
                    "format": "float64"
                },
                "tradeMetrics": {
                    "$ref": "#/definitions/domain.TradeMetrics"
                },
                "unrealizedProfit": {
                    "type": "number",
                    "format": "float64"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
...
    }
}